Remove duplicate export statements from adjust, and fix JSdocs
