+++ src/actions/adjust/OpacityAdjustAction.ts
@@ -4,7 +4,8 @@ import {Qualifier} from "../../internal/qualifier/Qualifier";
 /**
  * @description OpacityAction class, used to define the opacity of an asset
  * @class OpacityAdjustAction
- * @augments Action
+ * @memberOf Actions.Adjust
+ * @augments {SDK.Action}
  */
 class OpacityAdjustAction extends Action {
   constructor(level: number) {
@@ -13,4 +14,4 @@ class OpacityAdjustAction extends Action {
   }
 }
 
-export default OpacityAdjustAction;
+export {OpacityAdjustAction};
