+++ src/actions/adjust/simple/ViesusCorrectAdjustAction.ts
@@ -4,6 +4,7 @@ import {Qualifier} from "../../../internal/qualifier/Qualifier";
 /**
  * Enhances an image to its best visual quality with the Viesus Automatic Image Enhancement add-on.</br>
  * <b>Learn more:</b> {@link https://cloudinary.com/documentation/viesus_automatic_image_enhancement_addon | Viesus Automatic Image Enhancement.}
+ * @memberOf Actions.Adjust
  */
 class ViesusCorrectAdjustAction extends Action {
   private _noRedEye : boolean;
@@ -47,14 +48,4 @@ class ViesusCorrectAdjustAction extends Action {
   }
 }
 
-/**
- * @memberOf Actions.Adjust
- * @description
- * Enhances an image to its best visual quality with the Viesus Automatic Image Enhancement add-on.</br>
- * <b>Learn more:</b> {@link https://cloudinary.com/documentation/viesus_automatic_image_enhancement_addon | Viesus Automatic Image Enhancement.}
- */
-function viesusCorrect(): ViesusCorrectAdjustAction {
-  return new ViesusCorrectAdjustAction();
-}
-
-export default viesusCorrect;
+export {ViesusCorrectAdjustAction};
