--- src/actions/adjust/withLevel/opacityThreshold.ts
@@ -1,20 +1,0 @@
-import {EffectActionWithLevel}from "../../effect/EffectActions/EffectActionWithLevel";
-
-/**
- * @memberOf Actions.Adjust
- * @description  Causes all semi-transparent pixels in an image to be either fully transparent or fully opaque.
- *
- * Each pixel with an opacity lower than the specified threshold value is set to an opacity of 0%. Each pixel with
- * an opacity greater than the specified threshold is set to an opacity of 100%. For example, setting
- * opacity_threshold:0 makes all pixels non-transparent, while opacity_threshold:100 makes all semi-transparent
- * pixels fully transparent. Note: This effect can be a useful solution when PhotoShop PSD files are delivered in a
- * format supporting partial transparency, such as PNG, and the results without this effect are not as expected.
- *
- * @param {number} level The level of the threshold. (Range: 1 to 100, Server default: 50)
- * @return {EffectActionWithLevel}
- */
-function opacityThreshold(level?:number): EffectActionWithLevel {
-  return new EffectActionWithLevel('opacity_threshold', level);
-}
-
-export default opacityThreshold;
