+++ src/actions/effect/EffectActions/EffectActionWithBlendPercentage.ts
@@ -3,7 +3,8 @@ import {LeveledEffectAction} from "./LeveledEffectAction";
 /**
  * @description A class for all effects that include a blendPercentage method
  * @class EffectActionWithBlendPercentage
- * @augments LeveledEffectAction
+ * @memberOf Actions.Effect
+ * @extends {Actions.Effect.LeveledEffectAction}
  */
 class EffectActionWithBlendPercentage extends LeveledEffectAction {
   blendPercentage(value: number | string): this {
