+++ src/actions/effect/EffectActions/EffectActionWithStrength.ts
@@ -3,7 +3,8 @@ import {LeveledEffectAction} from "./LeveledEffectAction";
 /**
  * @description A class for all effects that include a strength method
  * @class EffectActionWithStrength
- * @augments LeveledEffectAction
+ * @memberOf Actions.Effect
+ * @extends {Actions.Effect.LeveledEffectAction}
  */
 class EffectActionWithStrength extends LeveledEffectAction {
   strength(value: number | string): this {
